ir_attachment_azure/requirements.txt
python-slugify<=5.0.0
